api_server/requirements.txt
Flask>=2.3.0
pandas>=2.0.0
selenium>=4.11.2
webdriver-manager>=3.8.6
beautifulsoup4>=4.12.2
langchain>=0.1.16
langchain-community>=0.0.32
langchain-core>=0.1.50
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.6
langchain-openai>=0.1.3
duckduckgo-search>=4.1
python-dotenv>=1.0.1
googlesearch-python>=1.1
torch>=2.2.0
torchvision>=0.17.0
--extra-index-url https://download.pytorch.org/whl/cu122
matplotlib>=3.8.0
numpy>=1.25.0
tqdm>=4.66.0
PyMuPDF>=1.23.21
pillow>=10.2.0
pytesseract>=0.3.10
sentence-transformers>=2.2.2
psycopg2-binary>=2.9.9
python-docx 
python-pptx
openpyxl 
xlrd
unstructured[all-docs]>=0.18.15
docling[full]
mlx-vlm
minio